lots of int->size_t conversions
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@22147 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ qfits-an/src/qfits_memory.c
@@ -449,7 +449,7 @@ void * qfits_memory_calloc(
 
 // copied from ioutils.c
 
-static void get_mmap_size(int start, int size, off_t* mapstart, size_t* mapsize, int* pgap) {
+static void get_mmap_size(size_t start, size_t size, off_t* mapstart, size_t* mapsize, int* pgap) {
 	int ps = getpagesize();
 	int gap = start % ps;
 	// start must be a multiple of pagesize.
